Guns in 51 US States Analysis(Stata)
The main objective of our project is to analyze the data to gain insights about the understanding of how shall-issues laws affect crime rate across 51 states in U.S from 1977 to 1999. The total number of observations in data set are 1173 which is spread across 51 states within 23 years. Ran pooled OLS, Fixed effect, Random effect,
Entity and Time Fixed Effect models to check for Heteroscedasticity, Endogeneity and collinearity problems in data and their effects.